customer-bot/requirements.txt
fastapi
uvicorn[standard]
pydantic
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pypdf>=3.12.0
openai>=1.0.0
tqdm
python-multipart
